refactor: move CostInfoToMetaMap to base package; drop duplicate stt copy
