+++ runtime/tts/cost.go
@@ -56,28 +56,10 @@ func ComputeTTSCost(svc any, text string, latency time.Duration) *types.CostInfo
 	return info
 }
 
-// CostInfoToMetaMap serializes a CostInfo into the map[string]any shape
-// that addCostFromMetaKey (and the arena statestore) expect when reading
-// ancillary cost from Message.Meta. The keys and types must match exactly
-// what addCostFromMetaKey reads (see cost_aggregation.go).
+// CostInfoToMetaMap is kept here as a deprecated alias for back-compat with
+// existing call sites. Prefer base.CostInfoToMetaMap directly.
+//
+// Deprecated: use base.CostInfoToMetaMap.
 func CostInfoToMetaMap(ci *types.CostInfo) map[string]any {
-	if ci == nil {
-		return nil
-	}
-	m := map[string]any{
-		"total_cost_usd":  ci.TotalCost,
-		"input_cost_usd":  ci.InputCostUSD,
-		"output_cost_usd": ci.OutputCostUSD,
-		"input_tokens":    ci.InputTokens,
-		"output_tokens":   ci.OutputTokens,
-		"capability":      ci.Capability,
-		"provider_name":   ci.ProviderName,
-	}
-	if len(ci.Quantities) > 0 {
-		m["quantities"] = ci.Quantities
-	}
-	if len(ci.DimensionMatch) > 0 {
-		m["dimension_match"] = ci.DimensionMatch
-	}
-	return m
+	return base.CostInfoToMetaMap(ci)
 }
